requirements.txt
streamlit
chromadb
sentence-transformers
transformers
google-generativeai
pypdf
tqdm
langdetect
speechrecognition
pydub
numpy
python-multipart
python-dotenv
streamlit-mic-recorder
Pillow
pytesseract
fpdf
